refactor: rename registerThunk to createThunk with deprecation shim (closes #175)
