+++ .gitignore
@@ -1,2 +1,3 @@
 /dist
 /coverage
+/node_modules
