+++ types/createThunk.d.ts
@@ -1,0 +1,4 @@
+import type { State } from './stateTypes';
+import type { ThunkModule } from './thunk';
+export default function createThunk<S extends State>(theDo: ThunkModule<S>): void;
+export declare function registerThunk<S extends State>(theDo: ThunkModule<S>): void;
