+++ types/index.d.ts
@@ -3,7 +3,7 @@ import type { Dispatch } from './dispatch';
 import type { DispatchFuncMap } from './dispatchFuncMap';
 import { genUUID } from './genUUID';
 import { type InitParams, init } from './init';
-import registerThunk from './registerThunk';
+import createThunk, { registerThunk } from './createThunk';
 import { remove } from './remove';
 import { setDefaultID } from './setDefaultID';
 import { getDefaultID, getNode, getState, mustGetState, mustGetStateByThunk } from './states';
@@ -12,4 +12,4 @@ import ThunkContext from './ThunkContext';
 import type { ThunkModule, ThunkModuleToFunc } from './thunk';
 import { setData, update } from './update';
 import useThunk, { type UseThunk } from './useThunk';
-export { registerThunk, useThunk, ThunkContext, type UseThunk, type State, type NodeState, type NodeStateMap, type ClassState, type GetClassState, type Thunk, type ThunkModule, type ThunkModuleToFunc, type Dispatch, type DispatchFuncMap, getDefaultID, getNode, getState, mustGetState, mustGetStateByThunk, init, type InitParams, setData, update, remove, setDefaultID, genUUID, };
+export { createThunk, registerThunk, useThunk, ThunkContext, type UseThunk, type State, type NodeState, type NodeStateMap, type ClassState, type GetClassState, type Thunk, type ThunkModule, type ThunkModuleToFunc, type Dispatch, type DispatchFuncMap, getDefaultID, getNode, getState, mustGetState, mustGetStateByThunk, init, type InitParams, setData, update, remove, setDefaultID, genUUID, };
